refactor: improvements
removed unused code
