+++ app/page.tsx
@@ -1,18 +1,16 @@
-import HeroSection, { HeroHighlightDemo } from "@/components/HeroSection";
+import HeroSection from "@/components/HeroSection";
 import HeroSectionCta from "@/components/HeroSectionCta";
 import Footer from "@/components/Footer";
 import FeaturesSection from "@/components/FeaturesSection";
-import TestimonialSection from "@/components/TestimonialSection ";
+import TestimonialSection from "@/components/TestimonialSection";
 import { CtaSection } from "@/components/CtaSection";
 
 export default async function Home() {
   return (
-    <section className="flex-col items-center justify-center bg-background">
+    <section className="flex-col items-center justify-center">
       <HeroSection>
         <HeroSectionCta />
       </HeroSection>
-
-      {/* <HeroHighlightDemo /> */}
 
       <FeaturesSection />
