+++ components/HeroSection.tsx
@@ -5,6 +5,7 @@ import Image from "next/image";
 import { useTheme } from "next-themes";
 import LightDash from "@/public/dash-light.png";
 import DarkDash from "@/public/dash-dark.png";
+import { Highlight } from "./ui/hero-highlight";
 
 export default function HeroSection({
   children,
@@ -53,32 +54,5 @@ export default function HeroSection({
         </div>
       </motion.div>
     </div>
-  );
-}
-
-import { HeroHighlight, Highlight } from "@/components/ui/hero-highlight";
-
-export function HeroHighlightDemo() {
-  return (
-    <HeroHighlight>
-      <motion.h1
-        initial={{
-          opacity: 0,
-          y: 20,
-        }}
-        animate={{
-          opacity: 1,
-          y: [20, -5, 0],
-        }}
-        transition={{
-          duration: 0.5,
-          ease: [0.4, 0.0, 0.2, 1],
-        }}
-        className="text-2xl px-4 md:text-4xl lg:text-5xl font-bold text-neutral-700 dark:text-white max-w-4xl leading-relaxed lg:leading-snug text-center mx-auto "
-      >
-        Master Your Notes <br />
-      </motion.h1>
-      <Highlight className="text-black dark:text-white">With Ease</Highlight>
-    </HeroHighlight>
   );
 }
